refactor: move service factory functions to server module
- Move create_service_factory, create_shared_service_instance, and get_rpyc_config from parameters.py to server.py
- Update imports in __init__.py and example code
- Add documentation for service factories and parameter handling in README
- Improve code organization with better separation of concerns
