+++ src/dcc_mcp_rpyc/parameters.py
@@ -6,7 +6,7 @@ with RPyC-specific functionality to handle remote procedure calls.
 
 # Import built-in modules
 import logging
-from typing import Any, Dict, List, Optional, Tuple, Union, Type, Callable
+from typing import Any, Dict, List, Optional, Tuple, Union
 
 # Import third-party modules
 import rpyc
@@ -14,7 +14,6 @@ from rpyc.utils import classic
 
 # Import local modules
 from dcc_mcp_core.parameters.processor import process_parameters as core_process_parameters
-from dcc_mcp_core.parameters.processor import process_parameter_value
 
 logger = logging.getLogger(__name__)
 
@@ -96,125 +95,3 @@ def execute_remote_command(connection: rpyc.Connection, command: str, *args, **k
     
     # Execute the command with processed arguments
     return cmd(*processed_args, **processed_kwargs)
-
-
-def create_service_factory(service_class: Type[rpyc.Service], *args, **kwargs) -> Callable:
-    """Create a factory function for a service class with bound arguments.
-    
-    This is similar to rpyc.utils.helpers.classpartial but with more flexibility
-    for DCC-MCP specific needs. It allows creating a service factory that will
-    instantiate the service with the provided arguments for each new connection.
-    
-    Args:
-        service_class: The RPyC service class to create a factory for
-        *args: Positional arguments to pass to the service constructor
-        **kwargs: Keyword arguments to pass to the service constructor
-        
-    Returns:
-        A factory function that creates service instances
-    """
-    class ServiceFactory:
-        """Factory for creating service instances with bound arguments."""
-        
-        def __init__(self):
-            self.service_class = service_class
-            self.args = args
-            self.kwargs = kwargs
-        
-        def get_service_name(self):
-            """Get the service name from the service class."""
-            if hasattr(service_class, 'get_service_name'):
-                return service_class.get_service_name()
-            # Default behavior: use class name without 'Service' suffix
-            name = service_class.__name__
-            if name.endswith("Service"):
-                name = name[:-7]
-            return name.upper()
-        
-        def get_service_aliases(self):
-            """Get the service aliases from the service class."""
-            if hasattr(service_class, 'get_service_aliases'):
-                return service_class.get_service_aliases()
-            return (self.get_service_name(),)
-        
-        def __call__(self, conn):
-            """Create a new service instance for each connection."""
-            instance = service_class(*args, **kwargs)
-            if hasattr(instance, 'on_connect'):
-                instance.on_connect(conn)
-            return instance
-    
-    return ServiceFactory()
-
-
-def create_shared_service_instance(service_class: Type[rpyc.Service], *args, **kwargs) -> rpyc.Service:
-    """Create a shared service instance that will be used for all connections.
-    
-    This function creates a single instance of the service that will be
-    shared among all connections. This is useful when you want to share
-    state between different connections.
-    
-    Args:
-        service_class: The RPyC service class to instantiate
-        *args: Positional arguments to pass to the service constructor
-        **kwargs: Keyword arguments to pass to the service constructor
-        
-    Returns:
-        A service instance that will be shared among all connections
-    """
-    instance = service_class(*args, **kwargs)
-    
-    # Create a wrapper that handles on_connect
-    class SharedServiceWrapper:
-        """Wrapper for a shared service instance."""
-        
-        def __init__(self, instance):
-            self.instance = instance
-        
-        def get_service_name(self):
-            """Get the service name from the instance."""
-            if hasattr(instance, 'get_service_name'):
-                return instance.get_service_name()
-            # Default behavior
-            name = service_class.__name__
-            if name.endswith("Service"):
-                name = name[:-7]
-            return name.upper()
-        
-        def get_service_aliases(self):
-            """Get the service aliases from the instance."""
-            if hasattr(instance, 'get_service_aliases'):
-                return instance.get_service_aliases()
-            return (self.get_service_name(),)
-        
-        def __call__(self, conn):
-            """Return the shared instance for each connection."""
-            if hasattr(instance, 'on_connect'):
-                instance.on_connect(conn)
-            return instance
-    
-    return SharedServiceWrapper(instance)
-
-
-def get_rpyc_config(allow_all_attrs=False, allow_public_attrs=True, allow_pickle=False):
-    """Get a configuration dictionary for RPyC connections.
-    
-    This function creates a configuration dictionary with common settings
-    for RPyC connections in the DCC-MCP ecosystem.
-    
-    Args:
-        allow_all_attrs: Whether to allow access to all attributes
-        allow_public_attrs: Whether to allow access to public attributes
-        allow_pickle: Whether to allow pickle serialization
-        
-    Returns:
-        A configuration dictionary for RPyC connections
-    """
-    return {
-        'allow_all_attrs': allow_all_attrs,
-        'allow_public_attrs': allow_public_attrs,
-        'allow_pickle': allow_pickle,
-        'sync_request_timeout': 30,
-        'allow_safe_attrs': True,
-        'allow_exposed_attrs': True,
-    }
