+++ src/dcc_mcp_rpyc/server.py
@@ -7,19 +7,13 @@ DCC software environments, with specific handling for different DCC threading mo
 # Import built-in modules
 from abc import ABC
 from abc import abstractmethod
+from typing import Any, Callable, Dict, List, Optional, Type, Tuple, Union
 import functools
 import logging
 import os
+import socket
 import threading
 import time
-from typing import Any
-from typing import Callable
-from typing import Dict
-from typing import Optional
-from typing import Type
-from typing import TypeVar
-from typing import Union
-from typing import cast
 
 # Import third-party modules
 import rpyc
@@ -591,9 +585,134 @@ def create_dcc_server(
     )
 
 
+def create_service_factory(service_class: Type[rpyc.Service], *args, **kwargs) -> Callable:
+    """Create a factory function for a service class with bound arguments.
+    
+    This is similar to rpyc.utils.helpers.classpartial but with more flexibility
+    for DCC-MCP specific needs. It allows creating a service factory that will
+    instantiate the service with the provided arguments for each new connection.
+    
+    Args:
+        service_class: The RPyC service class to create a factory for
+        *args: Positional arguments to pass to the service constructor
+        **kwargs: Keyword arguments to pass to the service constructor
+        
+    Returns:
+        A factory function that creates service instances
+    """
+    class ServiceFactory:
+        """Factory for creating service instances with bound arguments."""
+        
+        def __init__(self):
+            self.service_class = service_class
+            self.args = args
+            self.kwargs = kwargs
+        
+        def get_service_name(self):
+            """Get the service name from the service class."""
+            if hasattr(service_class, 'get_service_name'):
+                return service_class.get_service_name()
+            # Default behavior: use class name without 'Service' suffix
+            name = service_class.__name__
+            if name.endswith("Service"):
+                name = name[:-7]
+            return name.upper()
+        
+        def get_service_aliases(self):
+            """Get the service aliases from the service class."""
+            if hasattr(service_class, 'get_service_aliases'):
+                return service_class.get_service_aliases()
+            return (self.get_service_name(),)
+        
+        def __call__(self, conn):
+            """Create a new service instance for each connection."""
+            instance = service_class(*args, **kwargs)
+            if hasattr(instance, 'on_connect'):
+                instance.on_connect(conn)
+            return instance
+    
+    return ServiceFactory()
+
+
+def create_shared_service_instance(service_class: Type[rpyc.Service], *args, **kwargs) -> rpyc.Service:
+    """Create a shared service instance that will be used for all connections.
+    
+    This function creates a single instance of the service that will be
+    shared among all connections. This is useful when you want to share
+    state between different connections.
+    
+    Args:
+        service_class: The RPyC service class to instantiate
+        *args: Positional arguments to pass to the service constructor
+        **kwargs: Keyword arguments to pass to the service constructor
+        
+    Returns:
+        A service instance that will be shared among all connections
+    """
+    instance = service_class(*args, **kwargs)
+    
+    # Create a wrapper that handles on_connect
+    class SharedServiceWrapper:
+        """Wrapper for a shared service instance."""
+        
+        def __init__(self, instance):
+            self.instance = instance
+        
+        def get_service_name(self):
+            """Get the service name from the instance."""
+            if hasattr(instance, 'get_service_name'):
+                return instance.get_service_name()
+            # Default behavior
+            name = service_class.__name__
+            if name.endswith("Service"):
+                name = name[:-7]
+            return name.upper()
+        
+        def get_service_aliases(self):
+            """Get the service aliases from the instance."""
+            if hasattr(instance, 'get_service_aliases'):
+                return instance.get_service_aliases()
+            return (self.get_service_name(),)
+        
+        def __call__(self, conn):
+            """Return the shared instance for each connection."""
+            if hasattr(instance, 'on_connect'):
+                instance.on_connect(conn)
+            return instance
+    
+    return SharedServiceWrapper(instance)
+
+
+def get_rpyc_config(allow_all_attrs=False, allow_public_attrs=True, allow_pickle=False):
+    """Get a configuration dictionary for RPyC connections.
+    
+    This function creates a configuration dictionary with common settings
+    for RPyC connections in the DCC-MCP ecosystem.
+    
+    Args:
+        allow_all_attrs: Whether to allow access to all attributes
+        allow_public_attrs: Whether to allow access to public attributes
+        allow_pickle: Whether to allow pickle serialization
+        
+    Returns:
+        A configuration dictionary for RPyC connections
+    """
+    return {
+        'allow_all_attrs': allow_all_attrs,
+        'allow_public_attrs': allow_public_attrs,
+        'allow_pickle': allow_pickle,
+        'sync_request_timeout': 30,
+        'allow_safe_attrs': True,
+        'allow_exposed_attrs': True,
+    }
+
+
 __all__ = [
     "DCCRPyCService",
     "DCCServer",
     "create_dcc_server",
     "create_raw_threaded_server",
+    "create_service_factory",
+    "create_shared_service_instance",
+    "get_rpyc_config",
 ]
